Update README.md to remove need for Wix Site
People like going on github and reading the information of the project on github. Making them go to a seperate wix site is confusing and doesn't look as nice. The github readme can be updated with the screenshots and information as it changes and the Patreon can be used for people to read Blog Post style posts and support development